+++ vscode-smart-contracts/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Update icon of the extension to better support dark themes.
+- Show an error when unable to determine the project, when running build and test commands.
 ## 1.0.1
 
 - Initial release